Starting from scratch on Todo, not copying Vanilla!
I want to get the string to 'clear' from the first input field, and
re-appear in te new to do field as the "Label", between the checkbox
and the button.
This script will be triggered by the user ''entering'' the text.
I want this to repeat when the placeholder text in Input field is typed
over and "Entered' via keyboard action. I think the event is just
called .change
Next, I want to create a script that changes the checkmark from one
state to another, and then make the button hide until hover (mousein).
havin fun so far =]